requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
langchain-google-genai
